requirements.txt
discord~=1.0.1
python-dotenv~=0.15.0
PyNaCl~=1.4.0
requests~=2.25.0